MauiScan.ML.Training/requirements.txt
# PyTorch 深度学习框架
torch>=2.0.0
torchvision>=0.15.0

# ONNX 模型导出
onnx>=1.15.0
onnxruntime>=1.20.0

# 图像处理
opencv-python>=4.8.0
Pillow>=10.0.0

# 数据处理
numpy>=1.24.0

# 训练工具
tqdm>=4.65.0
tensorboard>=2.13.0

# 配置管理
pyyaml>=6.0
